Appending target framework to the OutputPath/IntermediateOutputPath
We only can do this when AndroidUseLatestPlatformSdk is false due to the
need of evaluating the MsBuild properties before calculating the latest SDK
using the _SetLatestTargetFrameworkVersion target (Xamarin.Android.Common.targets)

XA heavily leverages generated files in the intermediate output directory to achieve incremental build,
which is key since a lot of the Android/Java SDK tools are slow.

In XA, targeting different API levels is done by changing the TargetFramework property in the project.
This effectively invalidates all previous intermediate artifacts. Since this is not an uncommon
action (i.e. user is trying out various target API levels to see which one works best for their needs),
we'd like to preserve whatever artifacts were generated in a previous run for a given framework.

So rather than invalidating all incremental files when that property is changed (i.e. doing a full
clean+build on the next run), a better approach is to append the target framework to the intermediate output path,
following in the footsteps from .NETCore/.NETStandard (which always append the target framework too).